+++ src/pages/Flows/Components/select/form.tsx
@@ -1,12 +1,12 @@
-import { getContactDetail } from '@/services/contacts/api';
-import { genContactFromData } from '@/services/contacts/util';
+import UnitGroupFromMini from '@/pages/Unitgroups/Components/select/formMini';
+import { getFlowDetail } from '@/services/flows/api';
+import { genFlowFromData } from '@/services/flows/util';
 import { ProFormInstance } from '@ant-design/pro-components';
 import { Button, Card, Col, Divider, Form, Input, Row, Space } from 'antd';
 import React, { FC, ReactNode, useEffect, useState } from 'react';
 import { FormattedMessage } from 'umi';
-import ContactView from '../view';
-import ContactSelectDrawer from './drawer';
-
+import FlowsView from '../view';
+import FlowsSelectDrawer from './drawer';
 const { TextArea } = Input;
 
 type Props = {
@@ -14,59 +14,52 @@ type Props = {
   label: ReactNode | string;
   lang: string;
   formRef: React.MutableRefObject<ProFormInstance | undefined>;
+  drawerVisible: boolean;
   onData: () => void;
 };
 
-const ContactSelectFrom: FC<Props> = ({ name, label, lang, formRef, onData }) => {
+const FlowsSelectForm: FC<Props> = ({ name, label, lang, formRef, drawerVisible, onData }) => {
   const [id, setId] = useState<string | undefined>(undefined);
 
-  const handletContactData = (rowKey: any) => {
-    getContactDetail(rowKey).then(async (result: any) => {
-      const selectedData = genContactFromData(result.data?.json?.contactDataSet ?? {});
+  const handletFlowsData = (rowKey: any) => {
+    getFlowDetail(rowKey).then(async (result: any) => {
+      const selectedData = genFlowFromData(result.data?.json?.flowDataSet ?? {});
       await formRef.current?.setFieldValue(name, {
         '@refObjectId': `${rowKey}`,
-        '@type': 'contact data set',
-        '@uri': `../contacts/${rowKey}.xml`,
+        '@type': 'flow data set',
+        '@uri': `../flows/${rowKey}.xml`,
         'common:shortDescription':
-          selectedData?.contactInformation?.dataSetInformation?.['common:shortName']?.map(
-            (item: any) => {
-              return {
-                ...item,
-                '#text': `${item['#text']}, ${
-                  selectedData?.contactInformation?.dataSetInformation?.email ?? ''
-                }`,
-              };
-            },
-          ) ?? [],
-        '@version':
-          result.data.json?.contactDataSet?.administrativeInformation?.publicationAndOwnership?.[
-            'common:dataSetVersion'
-          ],
+          selectedData?.flowInformation?.dataSetInformation?.name?.baseName ?? [],
       });
       onData();
     });
   };
 
-  // const actionRef = React.useRef<ActionType | undefined>(undefined);
-
-  // const id = formRef.current?.getFieldValue([...name, '@refObjectId']);
-
   useEffect(() => {
-    setId(formRef.current?.getFieldValue([...name, '@refObjectId']));
+    if (drawerVisible) {
+      if (formRef.current?.getFieldValue([...name, '@refObjectId'])) {
+        setId(formRef.current?.getFieldValue([...name, '@refObjectId']));
+      }
+    }
   });
 
   return (
     <Card size="small" title={label}>
       <Space direction="horizontal">
         <Form.Item
-          label={<FormattedMessage id="pages.contact.refObjectId" defaultMessage="Ref Object Id" />}
+          label={
+            <FormattedMessage
+              id="pages.process.view.exchange.refObjectId"
+              defaultMessage="Ref Object Id"
+            />
+          }
           name={[...name, '@refObjectId']}
         >
           <Input disabled={true} style={{ width: '350px', color: '#000' }} />
         </Form.Item>
         <Space direction="horizontal" style={{ marginTop: '6px' }}>
-          <ContactSelectDrawer buttonType="text" lang={lang} onData={handletContactData} />
-          {id && <ContactView lang={lang} id={id} buttonType="text" />}
+          <FlowsSelectDrawer buttonType="text" lang={lang} onData={handletFlowsData} />
+          {id && <FlowsView lang={lang} id={id} buttonType="text" />}
           {id && (
             <Button
               onClick={() => {
@@ -80,25 +73,22 @@ const ContactSelectFrom: FC<Props> = ({ name, label, lang, formRef, onData }) =>
         </Space>
       </Space>
       <Form.Item
-        label={<FormattedMessage id="pages.contact.type" defaultMessage="Type" />}
+        label={<FormattedMessage id="pages.process.view.exchange.type" defaultMessage="Type" />}
         name={[...name, '@type']}
       >
         <Input disabled={true} style={{ color: '#000' }} />
       </Form.Item>
       <Form.Item
-        label={<FormattedMessage id="pages.contact.uri" defaultMessage="URI" />}
+        label={<FormattedMessage id="pages.process.view.exchange.uri" defaultMessage="URI" />}
         name={[...name, '@uri']}
       >
         <Input disabled={true} style={{ color: '#000' }} />
       </Form.Item>
-      <Form.Item
-        label={<FormattedMessage id="pages.contact.version" defaultMessage="Version" />}
-        name={[...name, '@version']}
-      >
-        <Input disabled={true} style={{ color: '#000' }} />
-      </Form.Item>
       <Divider orientationMargin="0" orientation="left" plain>
-        <FormattedMessage id="pages.contact.shortDescription" defaultMessage="Short Description" />
+        <FormattedMessage
+          id="pages.process.view.exchange.shortDescription"
+          defaultMessage="Short Description"
+        />
       </Divider>
       <Form.Item>
         <Form.List name={[...name, 'common:shortDescription']}>
@@ -128,8 +118,16 @@ const ContactSelectFrom: FC<Props> = ({ name, label, lang, formRef, onData }) =>
           )}
         </Form.List>
       </Form.Item>
+
+      <UnitGroupFromMini
+        id={id}
+        idType={'flow'}
+        name={name}
+        formRef={formRef}
+        drawerVisible={drawerVisible}
+      />
     </Card>
   );
 };
 
-export default ContactSelectFrom;
+export default FlowsSelectForm;
