+++ src/pages/Processes/Components/Exchange/edit.tsx
@@ -1,5 +1,5 @@
-import LangTextItemFrom from '@/components/LangTextItem/from';
-import FlowsSelectFrom from '@/pages/Flows/Components/select/from';
+import LangTextItemForm from '@/components/LangTextItem/form';
+import FlowsSelectForm from '@/pages/Flows/Components/select/form';
 import styles from '@/style/custom.less';
 import { CloseOutlined, FormOutlined } from '@ant-design/icons';
 import { ActionType } from '@ant-design/pro-components';
@@ -166,7 +166,7 @@ const ProcessExchangeEdit: FC<Props> = ({
                 ]}
               />
             </Form.Item>
-            <FlowsSelectFrom
+            <FlowsSelectForm
               name={['referenceToFlowDataSet']}
               label={
                 <FormattedMessage
@@ -218,7 +218,7 @@ const ProcessExchangeEdit: FC<Props> = ({
                 defaultMessage="General Comment"
               />
             </Divider>
-            <LangTextItemFrom name="generalComment" label="General Comment" />
+            <LangTextItemForm name="generalComment" label="General Comment" />
 
             <Card
               size="small"
@@ -248,7 +248,7 @@ const ProcessExchangeEdit: FC<Props> = ({
                       defaultMessage="Functional Unit Or Other"
                     />
                   </Divider>
-                  <LangTextItemFrom name="functionalUnitOrOther" label="Functional Unit Or Other" />
+                  <LangTextItemForm name="functionalUnitOrOther" label="Functional Unit Or Other" />
                 </>
               ) : (
                 <></>
