+++ src/pages/Sources/Components/select/form.tsx
@@ -1,13 +1,12 @@
-import UnitGroupFromMini from '@/pages/Unitgroups/Components/select/fromMini';
-import { getFlowpropertyDetail } from '@/services/flowproperties/api';
-import { genFlowpropertyFromData } from '@/services/flowproperties/util';
+import { getSourceDetail } from '@/services/sources/api';
+import { genSourceFromData } from '@/services/sources/util';
 import { ProFormInstance } from '@ant-design/pro-components';
+import { FormattedMessage } from '@umijs/max';
 import { Button, Card, Col, Divider, Form, Input, Row, Space } from 'antd';
 import React, { FC, ReactNode, useEffect, useState } from 'react';
-import { FormattedMessage } from 'umi';
-import FlowpropertyView from '../view';
-import FlowpropertiesSelectDrawer from './drawer';
-// import LangTextItemFrom from '@/components/LangTextItem/from';
+import SourceView from '../view';
+import SourceSelectDrawer from './drawer';
+
 const { TextArea } = Input;
 
 type Props = {
@@ -15,61 +14,44 @@ type Props = {
   label: ReactNode | string;
   lang: string;
   formRef: React.MutableRefObject<ProFormInstance | undefined>;
-  drawerVisible: boolean;
   onData: () => void;
 };
 
-const FlowpropertiesSelectFrom: FC<Props> = ({
-  name,
-  label,
-  lang,
-  formRef,
-  drawerVisible,
-  onData,
-}) => {
+const SourceSelectForm: FC<Props> = ({ name, label, lang, formRef, onData }) => {
   const [id, setId] = useState<string | undefined>(undefined);
 
-  const handletFlowpropertyData = (rowKey: any) => {
-    getFlowpropertyDetail(rowKey).then(async (result: any) => {
-      const selectedData = genFlowpropertyFromData(result.data?.json?.flowPropertyDataSet ?? {});
+  const handletSourceData = (rowKey: any) => {
+    getSourceDetail(rowKey).then(async (result: any) => {
+      const selectedData = genSourceFromData(result.data?.json?.sourceDataSet ?? {});
       await formRef.current?.setFieldValue(name, {
         '@refObjectId': `${rowKey}`,
-        '@type': 'flow property data set',
-        '@uri': `../flowproperties/${rowKey}.xml`,
+        '@type': 'source data set',
+        '@uri': `../sources/${rowKey}.xml`,
         'common:shortDescription':
-          selectedData?.flowPropertiesInformation?.dataSetInformation?.['common:name'] ?? [],
+          selectedData?.sourceInformation?.dataSetInformation?.['common:shortName'] ?? [],
       });
       onData();
     });
   };
 
+  // const id = formRef.current?.getFieldValue([...name, '@refObjectId']);
+
   useEffect(() => {
-    if (formRef.current?.getFieldValue([...name, '@refObjectId'])) {
-      setId(formRef.current?.getFieldValue([...name, '@refObjectId']));
-    }
+    setId(formRef.current?.getFieldValue([...name, '@refObjectId']));
   });
 
   return (
     <Card size="small" title={label}>
       <Space direction="horizontal">
         <Form.Item
-          label={
-            <FormattedMessage
-              id="pages.flow.view.flowProperties.refObjectId"
-              defaultMessage="Ref Object Id"
-            />
-          }
+          label={<FormattedMessage id="pages.contact.refObjectId" defaultMessage="Ref Object Id" />}
           name={[...name, '@refObjectId']}
         >
           <Input disabled={true} style={{ width: '350px', color: '#000' }} />
         </Form.Item>
         <Space direction="horizontal" style={{ marginTop: '6px' }}>
-          <FlowpropertiesSelectDrawer
-            buttonType="text"
-            lang={lang}
-            onData={handletFlowpropertyData}
-          />
-          {id && <FlowpropertyView lang={lang} id={id} buttonType="text" />}
+          <SourceSelectDrawer buttonType="text" lang={lang} onData={handletSourceData} />
+          {id && <SourceView lang={lang} id={id} buttonType="text" />}
           {id && (
             <Button
               onClick={() => {
@@ -83,30 +65,20 @@ const FlowpropertiesSelectFrom: FC<Props> = ({
         </Space>
       </Space>
       <Form.Item
-        label={<FormattedMessage id="pages.flow.view.flowProperties.type" defaultMessage="Type" />}
+        label={<FormattedMessage id="pages.contact.type" defaultMessage="Type" />}
         name={[...name, '@type']}
       >
         <Input disabled={true} style={{ color: '#000' }} />
       </Form.Item>
       <Form.Item
-        label={<FormattedMessage id="pages.flow.view.flowProperties.uri" defaultMessage="URI" />}
+        label={<FormattedMessage id="pages.contact.uri" defaultMessage="URI" />}
         name={[...name, '@uri']}
       >
         <Input disabled={true} style={{ color: '#000' }} />
       </Form.Item>
-      {/* <Form.Item label="Version" name={[...name, '@version']}>
-        <Input disabled={true} />
-      </Form.Item> */}
       <Divider orientationMargin="0" orientation="left" plain>
-        <FormattedMessage
-          id="pages.flow.view.flowProperties.shortDescription"
-          defaultMessage="Short Description"
-        />
+        <FormattedMessage id="pages.contact.shortDescription" defaultMessage="Short Description" />
       </Divider>
-      {/* <LangTextItemFrom
-        name={[...name, 'common:shortDescription']}
-        label="Short Description"
-      /> */}
       <Form.Item>
         <Form.List name={[...name, 'common:shortDescription']}>
           {(subFields) => (
@@ -135,16 +107,8 @@ const FlowpropertiesSelectFrom: FC<Props> = ({
           )}
         </Form.List>
       </Form.Item>
-
-      <UnitGroupFromMini
-        id={id}
-        idType={'flowproperty'}
-        name={name}
-        formRef={formRef}
-        drawerVisible={drawerVisible}
-      />
     </Card>
   );
 };
 
-export default FlowpropertiesSelectFrom;
+export default SourceSelectForm;
