+++ src/services/unitgroups/api.ts
@@ -1,6 +1,5 @@
 import { supabase } from '@/services/supabase';
 import { SortOrder } from 'antd/lib/table/interface';
-import { v4 } from 'uuid';
 import {
   classificationToString,
   genClassificationZH,
@@ -13,7 +12,7 @@ import { genUnitGroupJsonOrdered } from './util';
 const table_name = 'unitgroups';
 
 export async function createUnitGroup(data: any) {
-  const newID = v4();
+  // const newID = v4();
   const oldData = {
     unitGroupDataSet: {
       '@xmlns': 'http://lca.jrc.it/ILCD/UnitGroup',
@@ -24,10 +23,10 @@ export async function createUnitGroup(data: any) {
         'http://lca.jrc.it/ILCD/UnitGroup ../../schemas/ILCD_UnitGroupDataSet.xsd',
     },
   };
-  const newData = genUnitGroupJsonOrdered(newID, data, oldData);
+  const newData = genUnitGroupJsonOrdered(data.id, data, oldData);
   const result = await supabase
     .from(table_name)
-    .insert([{ id: newID, json_ordered: newData }])
+    .insert([{ id: data.id, json_ordered: newData }])
     .select();
   return result;
 }
